jsonapi: validate upgrade ids before excuting, so that have no streamed response.
The problem is that when we start executing upgrades we have a streamed response.
When we have a streamed response, we can no longer return a JSON error response.
The problem cannot be properly tested since we do not have real streamed responses
in testing and thus we can change the response status, headers and content after
starting to "stream".