Fix post upgrade ordering which was broken since 1.11.0.
The post upgrade ordering was changed in #64 to be reversed.
The change in #64 was done out of confusion and did the wrong thing.

This changes it back to the right order:

Given the profiles:
- A -> B
- B -> C
- C -> -

The order should be:
C, B, A

This is important so that policies can change things done by its dependencies.