Update quickinstaller product version when upgrading a package.
When installing a product (or profile), quickinstaller stores its version.
When upgrading the package, the version is not updated because it is not reinstalled.

Since upgrading the package should update all things so that it is similar to the new
version of the package, we update the package version in quickinstaller to the current
product version.