remove "could not import blessed" warning.
The warning appears on stderr when running the bin/upgrade command.

The warning currently indicates that there was a problem with the Python
installation in use, but we aim to make "blessed" optional and thus no
longer consider this as an error / problem.

When merging stdout and stderr, such as it is done in tests, the warning
might get mixed up with json output, which is annoying.