Appease pyflakes.
Jenkins says in the pyflakes report: 'ftw' imported but unused.
If you use flake8, the 'noqa' hint helps,
but pyflakes in standalone mode is not impressed,
so we appease it by using (mentioning) the imported module.
Note that the '# pyflakes' comment has no technical effect.
it is merely there for human consumption,
to indicate why this line is there.