Implement inplace migrator.
The inplace migrator provides a fast and easy way for migrating content in
upgrade steps.
It can be used for example for migration from Archetypes to Dexterity.

The difference between Plone's standard migration and the inplace migration
is that the standard migration creates a new sibling and moves the children
and the inplace migration simply replaces the objects within the tree an
attaches the children to the new parent.
This is a much faster approach since no move / rename events are fired.