Migration: skip invalid relations.
It is possible to construct objects so that
`extract_relations` (`p.a.relationfield`) extracts UID-strings as
relation, instead of relation objects.

When feeding that to `_setRelation` (`z3c.relationfield`) for updating
the relation catalog, it does not understand it and fail with an
`AttributeError`.

We now skip those invalid relations and log a warning.