Avoid profile reinstall in upgrades.
Avoid coverriding customizations by reinstalling already installed
profiles while upgrading (Plone>=4.3.8).

If an Upgrade in `A` installes the new dependency `B`, which depends on
`C` but `C` was already installed and customized in `A` earlier,
we should not reinstall `C` so that we can avoid lossing `A`'s
customizations.

This is simply done by using the dependency strategy
`DEPENDENCY_STRATEGY_NEW`, which is available in
`Products.GenericSetup >= 1.8` (`Plone >= 4.3.8`).