fix bug when setting versions after upgrading.
In the version 2.3.0 we have introduced a check that avoids downgrading
when installing orphan upgrades.
Since the version comparison is made as string and not with parsing the
version, it will cause wrong results when switching versioning
systems (e.g. from 4020 to 20160101000000).
The effect will be that upgrades are proposed which should not be.