Fix encoding issues in ASCII terminals.
When bin/upgrade is ran in a terminal with LC_ALL=C, printed unicodes
will be converted to the terminal encoding (ASCII).
Since umlauts are not part of ASCII it will fail.

By already encoding unicodes to bytestrings we can awoid that print will
convert the strings. If the terminal really cannot print those
characters it will display placeholders instead.