Remove broken brains from catalog.
The upgrade step methods ``catalog_unrestricted_get_object``,
``catalog_unrestricted_search`` (with ``full_objects=True``) and
``objects`` will remove broken brains from the catalog.

When a brain's object cannot be loaded because it does no longer exist,
those methods will cause the catalog brain to be removed from the
catalog.

This will increase the changes that an upgrade will pass even when the
catalog is inconsistent. It will also help cleaning up inconsistent
state in the catalog on the go.

A brain is considered dead when a ``KeyError`` occurs while trying to
find the object by unrestricted-traversing the path of the brain.