Do not propose executed upgrades newer than current db version.
The goal of this change is that we do not want to propose upgrades
which we already know were executed but the current profile id is too
old.

With this change, upgrade steps are never proposed when the upgrade
step recorder tells that the upgrade step was already executed.
This is especially handy for customizations of the upgrade step
recording behavior.