Fix connection problem when zope.conf contains ip-address.
The command line tool extracts the zope port from the zope.conf.
The extraction was not robust enough and tripped when the optional
zope.conf setting "ip-address" was used.

This change fixes the problem and adds tests for various zope.conf
combinations.

The solution is to expect a whitespace before the address string,
which can either be a newline (no indentation) or a space (with
indentation).