Plone 4.3.7 tests: Pin previously unpinned p.a.contenttypes dependencies:
Especially the pinning for Products.DateRecurringIndex to 2.1 is needed to
get tests passing again for 4.3.7.

This is required because in versions newer than 2.1 Products.DateRecurringIndex
now depends on the 'ZODB' egg (instead of 'ZODB3') which results in a version
conflict around the 'transaction' module on Plone versions which don't use
ZODB > 3.x yet.