Tests: Choose actual port used by ZServer layer to run CommandAndInstance tests against:
The CommandAndInstance tests write a pseudo-zconf with the HTTP port
where they should expect a running Plone instance (as set up by the
plone.testing.z2.ZServer layer).

In more recent versions, if ZSERVER_PORT is not set, the ZServer layer
lets the OS pick a dynamic port for the ZServer it boots, instead of
defaulting to port 55001 (which it did earlier).

This meant that Plone 5 tests were failing locally, because we always
defaulted to 55001 if ZSERVER_PORT wasn't set (which it usually isn't when
running tests locally).

Therefore we instead just read out the actual port that the ZServer picked,
and write that to the pseudo-zconf.