Fix tests after Plone 4.3.19 release.
The tests were relying on a value in the attribute "rights", which was
not actually on the folder object but seems to have been inherited
through acquisition.
The Plone 4.3.19 release does no longer index parent attributes
through acquisition, thus the test was failing.

We have now a more relyable approach by removing an existing index and
readding it.