Added --allow-outdated option to install command.
This allows installing upgrades or profiles on a not up-to-date site.
Fixes issue #182.

I updated the `test_execute_upgrades_not_allowed_when_plone_outdated` method.
It was trying a command that failed with a 400 error because the Plone Site was outdated.
But when the Plone Site was *not* outdated, it would have failed with a different 400 error,
because the wanted upgrade step did not exist.
Updated it to use an existing upgrade step.
Duplicated the test to check that it works when adding our new `--allow-outdated` option.