Migrate creators even when dublin core behaviors are not enabled.
The Dexterity base class has support for creators (it implements
listCreators() and other methods.
The dublin core behavior makes it possible to edit the creators.

The migration did work when the dublin core behavior was enabled.
But when it was not enabled, the creators was overwritten by the
user executing the migration. This change updates the creators.