Add intermediate commit support to "Executioner".
We only support intermediate commits when installing upgrade profiles.
We take care to make sure the indexing queue is processed with logging
and that transaction notes are set up correctly.

The "portal_setup" version is now set after each upgrade is executed.
This ensures that upgrades will be resumed correctly when using
intermediate commits in case of failing upgrades.